udpated at

+++ models/private_bot_group_binding_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from models.base_model import BaseModel
@@ -9,7 +9,7 @@ class PrivateBotGroupBinding(BaseModel):
     __tablename__ = "private_bot_group_binding"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    private_chat_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    private_chat_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     bound_group_id: Mapped[int] = mapped_column(
         Integer, ForeignKey("chat_group.id"), nullable=False
     )
